% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{character_blanks_check}
\alias{character_blanks_check}
\title{Check whether all values are either character strings, blanks ("") or NA.}
\usage{
character_blanks_check(x)
}
\description{
To be used with \code{\link{col_check}}.
}

