% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_checks.R
\name{col_check}
\alias{col_check}
\title{Checking a single column for data fidelity}
\usage{
col_check(colname, data, fun, output = FALSE, stage = NULL)
}
\arguments{
\item{colname}{The name of the column within your dataframe.}

\item{data}{The dataframe containing the data.}

\item{fun}{The check function you'd like to apply to the data.}

\item{output}{Specifies whether you'd like an excel output file.
The excel file}

\item{fun}{The check function you'd like to apply to the data.}
}
\description{
Checking a single column for data fidelity
}

