% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Demonstration data created to resemble data collected from an educational
assessment.}
\format{A data frame with 20 rows and 5 variables: \describe{ \item{ID_var}{A
  student identification variable} \item{Fname}{First names of each student}
  \item{Var1}{One score} \item{Var2}{A second score} \item{Perf_Lvl}{Each
  student's performance level} \item{dates}{Birthdates}}}
\usage{
dataset
}
\description{
Demonstration data created to resemble data collected from an educational
assessment.
}
\keyword{datasets}

