% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{date_check}
\alias{date_check}
\title{Check whether all values fall within a date range. NA values are also
accepted.}
\usage{
date_check(x, begin, end, format = "\%m/\%d/\%Y")
}
\arguments{
\item{x}{the data that enters the function (the column specified in
\code{\link{col_check}})}

\item{begin}{the beginning acceptable date.}

\item{end}{the last acceptable date.}

\item{format}{a character string specifying the date format.}
}
\description{
To be used with \code{\link{col_check}}.
}
