% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{greater_than}
\alias{greater_than}
\title{Check whether values in column one are greater than their corresponding values
in the second column.}
\usage{
greater_than(col1, col2)
}
\arguments{
\item{col1}{the first column of data, specified in \code{\link{col_check}})}

\item{col2}{the second column of data, specified in \code{\link{col_check}})}
}
\description{
To be used with \code{\link{two_col_check}}.
}
