% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectr-package.R
\docType{package}
\name{inspectr-package}
\alias{inspectr-package}
\title{Storing, Manipulating and Analysis Spectroscopy and Associated Data in R}
\description{
Stores and eases the manipulation of spectra and associated data.
}
\author{
Pierre Roudier
}
