% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kenstone.R
\docType{methods}
\name{kenstone}
\alias{kenstone}
\alias{kenstone,Spectra-method}
\title{Kennard-Stone algorithm for optimal calibration set selection.}
\usage{
\S4method{kenstone}{Spectra}(x, size, ...)
}
\arguments{
\item{x}{a \code{Spectra} object}

\item{size}{a positive number, the number of items to choose from}
}
\value{
A vector of length size giving the indices of the selected
individuals in x.
}
\description{
An implemnentation of the Kennard-Stone algorithm for calibration set selection.
}
\author{
Pierre Roudier \url{pierre.roudier@gmail.com}
}
\references{
Kennard, L.A. Stone, Technometrics 11 (1969) 137.
}

