% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Spectra-methods.R
\docType{methods}
\name{nrow,Spectra-method}
\alias{nrow,Spectra-method}
\title{Returns the number of samples in the object}
\usage{
\S4method{nrow}{Spectra}(x)
}
\description{
Returns the number of samples in the object
}

