% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-methods.R
\docType{methods}
\name{plot_offset}
\alias{plot_offset}
\alias{plot_offset,Spectra-method}
\title{Offset plot of a collection of spectra}
\usage{
\S4method{plot_offset}{Spectra}(x, offset = 1)
}
\description{
Creates an offset plot of a collection of Spectra
}
\examples{
oz <- load_oz(3)
 plot_offset(oz)
 plot_offset(oz, 0.3)
 plot_offset(oz, 2)
}
\author{
Pierre Roudier
}

