% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-methods.R
\docType{methods}
\name{plot_stack}
\alias{plot_stack}
\title{Stacked plot of a collection of spectra}
\usage{
\S4method{plot_stack}{Spectra}(x)
}
\description{
Creates a stacked plot of a collection of Spectra
}
\examples{
oz <- load_oz(3)
 plot_stack(oz)
}
\author{
Pierre Roudier
}

