% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-methods.R
\docType{methods}
\name{plot_summary}
\alias{plot_summary}
\title{Summary plot of a collection of spectra}
\usage{
\S4method{plot_summary}{Spectra}(x, fun = mean, se = TRUE, ...)
}
\description{
Creates a summary plot of a collection of Spectra
}
\examples{
oz <- load_oz()
 plot_summary(oz)
}
\author{
Pierre Roudier
}

