% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/performance.R
\name{postResampleSpectro}
\alias{postResampleSpectro}
\alias{spectroSummary}
\title{Calculates performance indictors across resamples}
\usage{
spectroSummary(data, lev = NULL, model = NULL)
}
\description{
Given two numeric vectors of data, the root mean squared error, the R-squared, the bias, the RPD, the RPIQ, the CCC and the standard error are calculated. For two factors, the overall agreement rate and Kappa are determined.
}
\details{
This function extends \code{postResample} in the \code{caret} package.
}
\examples{
predicted <-  matrix(rnorm(50), ncol = 5)
observed <- rnorm(10)
apply(predicted, 2, postResampleSpectro, obs = observed)
}
\author{
Pierre Roudier, adapted from code from Max Kuhn
}

