% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Spectra-methods.R
\name{res.Spectra}
\alias{res.Spectra}
\title{Returns the spectral resolution of an object}
\usage{
\method{resolution}{Spectra}(x)
}
\arguments{
\item{object}{an object inheriting from \code{Spectra}}

\item{digits}{the number of significant digits}
}
\value{
a vector
}
\description{
Returns the spectral resolution of an object
}
\author{
Pierre Roudier \url{pierre.roudier@gmail.com}
}

