% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AAA-Classes.R
\docType{class}
\name{SpatialSpectra-class}
\alias{SpatialSpectra-class}
\title{Class for spectra collections that have
associated spatial point locations}
\description{
Class for spectra collections that have
associated spatial point locations
}

