% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AAA-Classes.R, R/SpatialSpectraDataFrame-methods.R
\docType{class}
\name{SpatialSpectraDataFrame-class}
\alias{SpatialSpectraDataFrame}
\alias{SpatialSpectraDataFrame-class}
\title{Class for spectra collections that have
associated attributes and spatial point locations}
\usage{
SpatialSpectraDataFrame(coords, nir, data, coords.nrs = numeric(0),
  proj4string = CRS(as.character(NA)), match.ID = TRUE, bbox = NULL,
  wl = numeric(), id = data.frame(id = NA), units = "nm")
}
\arguments{
\item{coords}{numeric matrix or data.frame with coordinates (each row is a
point)}

\item{nir}{a \code{"matrix"} or a \code{"data.frame"} object giving the
spectra values for each sample}

\item{data}{object of class \code{"data.frame"} containing the attribute
data}

\item{coords.nrs}{numeric; if present, records the column positions where in
\code{data} the coordinates were taken from (used by \link{coordinates<-})}

\item{proj4string}{projection string of class \link{CRS-class}}

\item{match.ID}{logical; if TRUE AND coords has rownames (i.e., coerced to a
matrix, \code{dimnames(coords)[[2]]} is not NULL), AND data has row.names
(i.e. is a data.frame), then the \code{SpatialSpectraDataFrame} object is
formed by matching the row names of both components, leaving the order of
the coordinates in tact. Checks are done to see whether both row names are
sufficiently unique, and all data are matched. If FALSE, coordinates and
data are simply "glued" together. If character: indicates the column in
\code{data} with coordinates IDs to match}

\item{bbox}{bounding box matrix, usually NULL and constructed from the data,
but may be passed through for coercion purposes if clearly needed}

\item{wl}{a numeric vector giving the wavelengths at with the spectra have
been measured}

\item{id}{a vector giving the unique id of each sample in the collection}

\item{units}{a character giving the unit in which the wavelengths values are
expressed}
}
\value{
a new \code{"SpatialSpectraDataFrame"} object
}
\description{
Class for spectra collections that have
associated attributes and spatial point locations

Constructor for the SpatialSpectraDataFrame class. Creates a
SpatialSpectraDataFrame object from scratch.
}
\section{Slots}{

\describe{
\item{\code{wl}}{object of class "\code{numeric}"; the wavelengths at which the spectra has been measured}

\item{\code{nir}}{object of class "\code{matrix}"; the spectra, with as many columns as wavelengths, and as many rows as samples}

\item{\code{id}}{object of class "\code{data.frame}" with one attribute; the identification strings for each sample in the collection}

\item{\code{units}}{object of class "\code{character}"; units in which the wavelengths are expressed}

\item{\code{data}}{object of class data.frame containing the attribute data (may or may not contain the coordinates in its columns)}

\item{\code{bbox}}{object of class "matrix"; bounding box}

\item{\code{proj4string}}{object of class "CRS"; projection string}

\item{\code{coords}}{object of class "matrix"; the coordinates matrix (points are rows in the matrix)}

\item{\code{coords.nrs}}{object of class logical; if TRUE, when the object was created the coordinates were retrieved from the data.frame, and hence stripped from it; after coercion to data.frame, e.g. by as.data.frame(x), coordinates will again be added (as first few columns) to the data.frame}
}}
\examples{
# Creating a SpatialSpectraDataFrame object from scratch
  my.wl <- 350:2500
  my.id <- LETTERS[1:4]
  my.nir <- matrix(runif(4*length(my.wl)), nrow=4)
  my.data <- data.frame(foo = runif(4), bar = LETTERS[1:4])
  my.coords <- matrix(runif(4), runif(4), ncol=2, nrow=4)

  my.ssdf <- SpatialSpectraDataFrame(wl = my.wl, nir = my.nir, id = my.id, data = my.data, coords = my.coords)
}
\author{
Pierre Roudier \email{pierre.roudier@gmail.com}

Pierre Roudier \url{pierre.roudier@gmail.com}
}
\seealso{
\code{\link{spectra}}, \code{\link{wl}}, \code{\link{coordinates}}, \code{\link{SpectraDataFrame-class}}

\code{\link{spectra}}, \code{\link{wl}},
\code{\link{SpatialSpectraDataFrame-class}}
}

