% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AAA-Classes.R
\docType{class}
\name{SpectraDataFrame-class}
\alias{SpectraDataFrame-class}
\title{Class for spectra collections with associated attributes}
\description{
Class for spectra collections with associated attributes
}
\section{Slots}{

\describe{
\item{\code{wl}}{object of class "\code{numeric}"; the wavelengths at which the spectra has been measured}

\item{\code{nir}}{object of class "\code{matrix}"; the spectra, with as many columns as wavelengths, and as many rows as samples}

\item{\code{id}}{object of class "\code{data.frame}" with one attribute; the identification strings for each sample in the collection}

\item{\code{units}}{object of class "\code{character}"; units in which the wavelengths are expressed}

\item{\code{data}}{object of class data.frame containing the attribute data}
}}
\author{
Pierre Roudier \email{pierre.roudier@gmail.com}
}
\seealso{
\code{\link{spectra}}, \code{\link{wl}}, \code{\link{Spectra-class}}
}

