% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/splice.R
\docType{methods}
\name{splice,Spectra-method}
\alias{"Splicing"}
\alias{and}
\alias{interpolation}
\alias{splice,Spectra-method}
\title{"Splicing" and interpolation of spectra}
\usage{
\S4method{splice}{Spectra}(x, wl = c(725:1020, 1801:1950),
  method = "spline")
}
\arguments{
\item{x}{a \code{Spectra} object}

\item{wl}{the wavelengths to cut out and interpolate}

\item{method}{the interpolation method. Available options are "linear",
"nearest", "pchip", "cubic", and "spline".}
}
\value{
an object of same class as x
}
\description{
This method mimicks the "splicing" method available in the ViewSpec Pro
software from ASD, which aims at correcting steps in the data.
}
\details{
It removes parts of the spectra defined by the \code{wl} vector, and
interpolates these parts using a method chosen using the \code{method}
option.

This function is a wrapper around \code{signal::interp1}.
}
\examples{
data(australia)
spectra(australia) <- sr_no ~ ... ~ 350:2500
oz_spliced <- splice(australia, wl = c(725:1020, 1801:1950), method = "spline")
plot(oz_spliced)
}
\author{
Pierre Roudier \url{pierre.roudier@gmail.com}
}

