% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Spectra-methods.R
\name{summary}
\alias{print.summary.Spectra}
\alias{summary}
\alias{summary.Spectra}
\title{Summary}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{an object of class \code{Spectra} or \code{SpectraDataFrame}}

\item{...}{dditional arguments passed to \code{summary}}
}
\value{
A \code{"summary.Spectra"} object
}
\description{
Summarize a Spectra* object.
}
\examples{
data(australia)
spectra(australia) <- sr_no ~ ... ~ 350:2500
summary(australia)
}
\author{
Pierre Roudier \url{pierre.roudier@gmail.com}
}

