% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Spectra-methods.R
\name{summary.Spectra}
\alias{print.summary.Spectra}
\alias{summary.Spectra}
\title{Summary}
\usage{
\method{summary}{Spectra}(object, ...)
}
\value{
A \code{"summary.Spectra"} object
}
\description{
Summarize a Spectra* object.
}
\examples{
data(australia)
spectra(australia) <- sr_no ~ ... ~ 350:2500
summary(australia)
}
\author{
Pierre Roudier \url{pierre.roudier@gmail.com}
}

