% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{val_check}
\alias{val_check}
\title{Check whether all values in the column fall within a set of user-defined
values.}
\usage{
val_check(x, values)
}
\arguments{
\item{x}{the data that enters the function (the column specified in
\code{\link{col_check}})}

\item{values}{contains a value or vector of values that contain the
acceptable value(s) that may be found in the column. These values may be
any data type - character strings, numeric values, etc.}
}
\description{
To be used with \code{\link{col_check}}.
}

