% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Spectra-methods.R
\docType{methods}
\name{wl_units}
\alias{wl_units}
\alias{wl_units,Spectra-method}
\alias{wl_units<-}
\alias{wl_units<-,Spectra-method}
\title{Wavelengths of Spectra* objects}
\usage{
\S4method{wl_units}{Spectra}(object)

\S4method{wl_units}{Spectra}(object) <- value
}
\arguments{
\item{object,x}{an object inheriting from class \code{Spectra}, or a
\code{"numeric"} vector for \code{resolution}}

\item{value}{a character string}
}
\value{
A vector
}
\description{
Retrieves the wavelengths units from
\code{Spectra*} objects.
}
\examples{
# Loading example data
data(australia)
spectra(australia) <- sr_no ~ ... ~ 350:2500

# Print wavelength information
wl(australia)
range(wl(australia))

# Manipulate wavelength information
wl(australia) <- wl(australia) + 1000
wl(australia)
}
\author{
Pierre Roudier \url{pierre.roudier@gmail.com}
}

