% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{character_blanks_check}
\alias{character_blanks_check}
\title{Check whether all values are either character strings. Blanks ("") and NA
values are not permitted.}
\usage{
character_blanks_check(x)
}
\arguments{
\item{x}{the data that enters the function (the column specified in
\code{\link{col_check}})}
}
\description{
To be used with \code{\link{col_check}}.
}

