% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{character_check}
\alias{character_check}
\title{Check whether all values are character strings or blanks (""). NA values are
not permitted.}
\usage{
character_check(x)
}
\arguments{
\item{x}{the data that enters the function (the column specified in
\code{\link{col_check}})}
}
\description{
To be used with \code{\link{col_check}}.
}

