% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_checks.R
\name{col_check}
\alias{col_check}
\title{Check a single column for data fidelity.}
\usage{
col_check(colname, data, fun, output = FALSE, loc = NULL, ...)
}
\arguments{
\item{colname}{character string specifying the name of the column within your
dataframe.}

\item{data}{the dataframe containing the data.}

\item{fun}{the check function you'd like to apply to the data.}

\item{output}{logical. If FALSE, the function returns a dataframe containing
only records that failed the specified check. If TRUE, invisibly prints an
excel output file containing only the records that failed the specified
check.}

\item{loc}{An optional character string that can be used to specify the desired 
location of the error output. Can only be used when output = TRUE. If output = TRUE
and loc is NULL, the error output gets put into the working directory.}

\item{...}{arguments to be passed through to the function specified in
\code{fun}}
}
\value{
col_check(output = FALSE) returns a dataframe containing only records
  that failed the specified check.

col_check(output = TRUE) invisibly prints an excel output file
  containing only records that failed the specified check.
}
\description{
To be used with user-defined functions or with built-in check functions such
as \code{\link{numeric_check}}, \code{\link{character_check}},
\code{\link{character_blanks_check}}, \code{\link{date_check}}, and
\code{\link{val_check}}.
}
\examples{
col_check(colname = "ID_var", data = dataset, fun = numeric_check,
output = TRUE)

col_check(colname = "FName", data = dataset, fun = character_check,
output = FALSE)

}
