% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{greater_than}
\alias{greater_than}
\title{Check whether values in column one are greater than their corresponding values
in the second column.}
\usage{
greater_than(col1, col2)
}
\description{
To be used with \code{\link{two_col_check}}.
}

