% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Spectra-methods.R, R/setters.R
\docType{methods}
\name{ids}
\alias{ids}
\alias{ids,Spectra-method}
\alias{ids<-}
\alias{ids<-,Spectra-method}
\alias{ids<-,SpectraDataFrame-method}
\title{Retrieves or sets the ids of a \code{Spectra*} object.}
\usage{
ids(object, ..., as.vector = TRUE)
ids(object) <- value

\S4method{ids}{Spectra}(object) <- value

\S4method{ids}{SpectraDataFrame}(object) <- value
}
\arguments{
\item{object}{an object of class \code{"Spectra"} or inheriting from this class}

\item{...}{currently ignored}

\item{as.vector}{should the IDs be returned as a vector or as a data.frame?}

\item{value}{character vector for new IDs}
}
\value{
The \code{ids} methods return a vector if \code{as.vector} is TRUE,
a \code{data.frame} otherwise. The \code{"ids<-"} method return a
\code{SpectraDataFrame} object (or a \code{Spectra} object if the column in
the data slot that has been used to initiate the IDs was the only
attribute).
}
\description{
Either retrieves the wavelengths from a \code{Spectra*} object, or creates a
\code{Spectra*} object from a \code{"data.frame"} object by setting some of
its columns as the wavelengths. The \code{"ids<-"} method for
\code{SpectraDataFrame} objects allows to use a formula interface to use a
column in its \code{data} slot as the object IDs (see the last example provided
in the Examples section).
}
\section{Methods}{
 \itemize{
 \item \code{ids(object, ..., as.vector = TRUE)}
 \item \code{ids(object) <- value}
}
}
\examples{
# Loading example data
data(australia)
spectra(australia) <- sr_no ~ ... ~ 350:2500

# Retrieving ids
ids(australia)

# Setting ids using a vector of values
ids(australia) <- seq_len(nrow(australia))
ids(australia)

# Setting ids using an attribute
australia$new_id <- seq_len(nrow(australia)) + 1000
ids(australia) <- ~ new_id
ids(australia)
}
\author{
Pierre Roudier \url{pierre.roudier@gmail.com}
}

