% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{less_than_equalto}
\alias{less_than_equalto}
\title{Check whether values in column one are less than or equal to their
corresponding values in the second column.}
\usage{
less_than_equalto(col1, col2)
}
\description{
To be used with \code{\link{two_col_check}}.
}

