% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Spectra-methods.R
\docType{methods}
\name{wl_units}
\alias{res}
\alias{res,Spectra-method}
\alias{res,integer-method}
\alias{res,numeric-method}
\alias{res.Spectra}
\alias{res.integer}
\alias{res.numeric}
\alias{wl_units}
\alias{wl_units,Spectra-method}
\alias{wl_units<-}
\alias{wl_units<-,Spectra-method}
\title{Wavelengths of Spectra* objects}
\usage{
\S4method{wl_units}{Spectra}(object)
}
\arguments{
\item{object,x}{an object inheriting from class \code{Spectra}, or a
\code{"numeric"} vector for \code{resolution}}

\item{value}{a character string}
}
\value{
A vector. \code{res} is giving NULL for irregularly spaced
wavelengths (eg when there is parts of the spectra data have been removed).
}
\description{
Retrieves the wavelengths units and the spectral resolution from
\code{Spectra*} objects.
}
\examples{
# Loading example data
data(australia)
spectra(australia) <- sr_no ~ ... ~ 350:2500

# Print wavelength information
wl(australia)
range(wl(australia))
res(australia)

# Manipulate wavelength information
wl(australia) <- wl(australia) + 1000
wl(australia)
}
\author{
Pierre Roudier \url{pierre.roudier@gmail.com}
}

